Fix incorrect merging of equal changes
If "mine" content is equal to "theirs", but different from "old" one,
diff3 -A still treats it as a conflict and prints "<<<<<<< old ======= theirs >>>>>>>"
instead of just skipping the hunk. Old version of this extension treated this
as if "mine" version was empty, which is incorrect.

It's impossible to get lines of "old" version from diff3 -E output,
so now patched wfDiff3() will remove these "equal conflicts" from the output manually.